Transform monolithic API into server actions
