+++ demos/it-request/webapp/src/app/actions/deleteWfRun.ts
@@ -1,0 +1,9 @@
+'use server';
+
+import { getClient } from '@/lib/lh-client';
+import type { WfRunId } from 'littlehorse-client/proto';
+
+export async function deleteWfRun(wfRunId: WfRunId): Promise<void> {
+  const client = getClient();
+  await client.deleteWfRun({ id: wfRunId });
+}
